bin/vpngate.sh: a little bugfix

+++ bin/vpngate.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
 confFile=vpngate.conf
-rm "${confFile}"
+[ -e "${confFile}" ] && rm "${confFile}"
 stat -c %n *.conf  || getvpngateconfs.sh
 find . -name '*.conf' -type f | tail -n 1 | xargs -I % mv -v % "${confFile}"
